Обновление структуры
